Problem 1 Sir!
I guess technically this is not all in one function or is it?  It jsut
like came to me easy tonight for some reason.  Googled a few things, had
some commands from the other day saved in the ol' Git.   I mean, it
isn't the PRETTIEST document, BUT i think it does what you ask, right?